+++ sdk/cosmos/azure-cosmos/azure/cosmos/_location_cache.py
@@ -134,6 +134,7 @@ class LocationCache(object):  # pylint: disable=too-many-public-methods,too-many
         refresh_time_interval_in_ms,
     ):
         self.preferred_locations = preferred_locations
+        self.effective_preferred_locations = []
         self.default_regional_endpoint = RegionalEndpoint(default_endpoint, default_endpoint)
         self.enable_endpoint_discovery = enable_endpoint_discovery
         self.use_multiple_write_locations = use_multiple_write_locations
@@ -245,7 +246,7 @@ class LocationCache(object):  # pylint: disable=too-many-public-methods,too-many
         return regional_endpoint.get_current()
 
     def should_refresh_endpoints(self):  # pylint: disable=too-many-return-statements
-        most_preferred_location = self.preferred_locations[0] if self.preferred_locations else None
+        most_preferred_location = self.effective_preferred_locations[0] if self.effective_preferred_locations else None
 
         # we should schedule refresh in background if we are unable to target the user's most preferredLocation.
         if self.enable_endpoint_discovery:
@@ -357,9 +358,6 @@ class LocationCache(object):  # pylint: disable=too-many-public-methods,too-many
         if refresh_cache:
             self.update_location_cache()
 
-    def get_preferred_locations(self):
-        return self.preferred_locations
-
     def update_location_cache(self, write_locations=None, read_locations=None, enable_multiple_writable_locations=None):
         if enable_multiple_writable_locations:
             self.enable_multiple_writable_locations = enable_multiple_writable_locations
@@ -387,6 +385,12 @@ class LocationCache(object):  # pylint: disable=too-many-public-methods,too-many
                     self.use_multiple_write_locations,
                 )
 
+        # if preferred locations is empty, we should use the read locations from gateway
+        if len(self.preferred_locations) == 0:
+            self.effective_preferred_locations = self.available_read_locations
+        else:
+            self.effective_preferred_locations = self.preferred_locations
+
         self.write_regional_endpoints = self.get_preferred_available_regional_endpoints(
             self.available_write_regional_endpoints_by_location,
             self.available_write_locations,
@@ -413,12 +417,12 @@ class LocationCache(object):  # pylint: disable=too-many-public-methods,too-many
                 or expected_available_operation == EndpointOperationType.ReadType
             ):
                 unavailable_endpoints = []
-                if self.preferred_locations:
+                if self.effective_preferred_locations:
                     # When client can not use multiple write locations, preferred locations
                     # list should only be used determining read endpoints order. If client
                     # can use multiple write locations, preferred locations list should be
                     # used for determining both read and write endpoints order.
-                    for location in self.preferred_locations:
+                    for location in self.effective_preferred_locations:
                         regional_endpoint = endpoints_by_location[location] if location in endpoints_by_location \
                             else None
                         if regional_endpoint: